+++ examples/with_health_extension.exs
@@ -17,7 +17,7 @@ NervesHubLinkAVM.start_link(
     "version" => "1.0.0",
     "platform" => "esp32"
   },
-  fwup_writer: MyApp.ESP32Writer,
+  firmware_writer: MyApp.ESP32Writer,
   client: MyApp.Client,
   extensions: [
     health: MyApp.HealthProvider